moved schema endpoint into one file
