+++ src/api/main.py
@@ -1,13 +1,11 @@
 from fastapi import FastAPI
 from src.logging.logging_setup import setup_logging
-from src.api.routers.schemas import component_schemas, job_schema
+from src.api.routers import schemas
 from src.api.helpers import autodiscover_components
 
 autodiscover_components("src.components")
 app = FastAPI()
-app.include_router(job_schema.router)
-app.include_router(component_schemas.router)
-
+app.include_router(schemas.router)
 
 if __name__ == "__main__":
     import uvicorn
